+++ packages/paraglide/paraglide-js-adapter-sveltekit/example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-js-adapter-sveltekit-example
 
+## 0.1.15
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-sveltekit@0.5.19
+
+## 0.1.14
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-sveltekit@0.5.18
+
 ## 0.1.13
 
 ### Patch Changes